renamed db model folder and setup cloudbuild stuff
