+++ .env
@@ -1,0 +1,3 @@
+FLASK_APP=foundation_api
+FLASK_ENV=development
+FLASK_HOST=${LOCAL_FLASK_HOST}
