--- .flaskenv
@@ -1,2 +1,0 @@
-FLASK_APP=foundation_api
-FLASK_ENV=development
