+++ deploy/prod_cloudbuild.yaml
@@ -1,0 +1,23 @@
+steps:
+- name: 'bash'
+  args: ['-c', './bash_scripts/set_env_vars.bash $$DB_USER_SECRET $$DB_PASSWORD_SECRET $$DB_HOST_SECRET $$DB_NAME_SECRET $$SENDGRID_API_KEY_SECRET ']
+  # args: ['bash_scripts/set_env_vars.bash $$DB_USER_SECRET $$DB_PASSWORD_SECRET $$DB_HOST_SECRET $$DB_NAME_SECRET $$SENDGRID_API_KEY_SECRET']
+  secretEnv: ['DB_PASSWORD_SECRET', 'DB_HOST_SECRET', 'DB_USER_SECRET', 'DB_NAME_SECRET', 'SENDGRID_API_KEY_SECRET']
+  
+- name: 'gcr.io/google.com/cloudsdktool/cloud-sdk'
+  entrypoint: 'bash'
+  args: ['-c', 'gcloud config set app/cloud_build_timeout 1600 && gcloud app deploy']
+timeout: '1600s'
+
+availableSecrets:
+  secretManager:
+  - versionName: projects/foundation-production/secrets/db_password/versions/latest
+    env: 'DB_PASSWORD_SECRET'
+  - versionName: projects/foundation-production/secrets/db_host/versions/latest
+    env: 'DB_HOST_SECRET'
+  - versionName: projects/foundation-production/secrets/db_user/versions/latest
+    env: 'DB_USER_SECRET'
+  - versionName: projects/foundation-production/secrets/db_name/versions/latest
+    env: 'DB_NAME_SECRET'
+  - versionName: projects/foundation-production/secrets/sendgrid-api-key/versions/latest
+    env: 'SENDGRID_API_KEY_SECRET'
