--- foundation_api/V1/mod_hello/models.py
@@ -1,1 +1,0 @@
-from foundation_api.db.model import Action
