+++ requirements.txt
@@ -39,6 +39,7 @@ pyasn1-modules==0.2.8
 PyJWT==2.0.1
 PyMySQL==1.0.2
 pyparsing==2.4.7
+python-dotenv==0.17.0
 pytz==2021.1
 PyYAML==5.4.1
 requests==2.25.1
